pyproject.toml
[project]
name = "awe"
version = "0.1.0"
description = "Agent Workflow Engine - A workflow engine for LLM agent orchestration"
readme = "README.md"
authors = [
    { name = "Noel", email = "acrecapv@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "aiosqlite>=0.20.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.0.0",
]

[project.scripts]
awe = "awe.cli:main"
awe-server = "awe.api.server:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
